requirements.txt
pyarrow
psycopg2-binary
pymongo
pandas
adbc-driver-postgresql
time
datetime
duckdb
numpy